requirements.txt
pyautogui>=0.9.53
pillow>=8.3.2
pynput>=1.7.6
pyperclip>=1.8.2
flask>=2.3.0
pywebview>=4.0.0